devops_incident_suite/requirements.txt
streamlit>=1.30.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.3.0
langchain-core>=0.3.0
langgraph>=0.2.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
